ACTIN-2158: refactoring
